maracker/requirements.txt
certifi==2017.4.17
chardet==3.0.3
Django==1.11.2
djangorestframework==3.6.3
idna==2.5
marathon==0.9.0
psycopg2==2.7.1
pytz==2017.2
PyYAML==3.12
requests==2.17.3
urllib3==1.21.1
